Revert "A very WIP directory restructure"
This reverts commit 0da2b2346bd07d8ca5a210dc872ee3fa8110b6bc.
