+++ src/main/kotlin/io/github/maheevil/modbot/App.kt
@@ -11,17 +11,17 @@ import com.kotlindiscord.kord.extensions.modules.extra.phishing.extPhishing
 import com.kotlindiscord.kord.extensions.utils.env
 import com.soywiz.korio.dynamic.KDynamic.Companion.toLong
 import dev.kord.common.entity.Snowflake
-import io.github.maheevil.modbot.extensions.moderation.RaidProt
-import io.github.maheevil.modbot.extensions.logging.LogEventListener
-import io.github.maheevil.modbot.extensions.moderation.ModerationCommands
+import io.github.maheevil.modbot.extensions.moderation.auto_moderation.RaidProt
+import io.github.maheevil.modbot.extensions.moderation.logging.LogEventListener
+import io.github.maheevil.modbot.extensions.moderation.util.ModerationCommands
 import io.github.maheevil.modbot.extensions.util.MiscCommands
-import io.github.maheevil.modbot.extensions.util.SuggestionCommand
-import io.github.maheevil.modbot.extensions.util.config.GuildConfigData
-import io.github.maheevil.modbot.extensions.util.config.deserializeAndLoadFromJson
+import io.github.maheevil.modbot.extensions.util.suggestions.SuggestionCommand
+import io.github.maheevil.modbot.util.config.GuildConfigData
+import io.github.maheevil.modbot.util.config.deserializeAndLoadFromJson
 
-var guildConfigDataMap = HashMap<Long, GuildConfigData>()
+var guildConfigDataMap = HashMap<Long,GuildConfigData>()
 
-val TEST_SERVER_ID = Snowflake(env("TEST_SERVER").toLong())
+//val TEST_SERVER_ID = Snowflake(env("TEST_SERVER").toLong())
 private val TOKEN = env("TOKEN")
 
 suspend fun main() {
@@ -35,7 +35,7 @@ suspend fun main() {
         }
         applicationCommands {
             enabled = true
-            defaultGuild(TEST_SERVER_ID)
+            //defaultGuild(TEST_SERVER_ID)
         }
         extensions {
             add(::ModerationCommands)
