+++ src/main/kotlin/io/github/maheevil/modbot/extensions/moderation/auto_moderation/RaidProt.kt
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-package io.github.maheevil.modbot.extensions.moderation
+package io.github.maheevil.modbot.extensions.moderation.auto_moderation
 
 import com.kotlindiscord.kord.extensions.extensions.Extension
 import com.kotlindiscord.kord.extensions.extensions.event
@@ -13,6 +13,7 @@ import dev.kord.core.entity.Guild
 import dev.kord.core.entity.channel.GuildMessageChannel
 import dev.kord.core.event.message.MessageCreateEvent
 import io.github.maheevil.modbot.extensions.moderation.logging.createAlertLog
+import io.github.maheevil.modbot.extensions.moderation.util.kickUserWithLog
 import io.github.maheevil.modbot.guildConfigDataMap
 
 class RaidProt : Extension() {
