+++ src/main/kotlin/io/github/maheevil/modbot/extensions/moderation/logging/LogCreator.kt
