+++ src/main/kotlin/io/github/maheevil/modbot/util/config/GuildConfigData.kt
@@ -3,7 +3,7 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-package io.github.maheevil.modbot.extensions.util.config
+package io.github.maheevil.modbot.util.config
 
 import dev.kord.common.entity.Snowflake
 import io.github.maheevil.modbot.guildConfigDataMap
